tests/wpt/web-platform-tests/feature-policy/reporting/xr-reporting.https.html

<!DOCTYPE html>
<html>
  <head>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
  </head>
  <body>
    <script>
var check_report_format = (reports, observer) => {
  let report = reports[0];
  assert_equals(report.type, "feature-policy-violation");
  assert_equals(report.url, document.location.href);
  assert_equals(report.body.featureId, "xr");
  assert_equals(report.body.sourceFile, document.location.href);
  assert_equals(typeof report.body.lineNumber, "number");
  assert_equals(typeof report.body.columnNumber, "number");
  assert_equals(report.body.disposition, "enforce");
};

promise_test(async (t) => {
  const report = new Promise(resolve => {
    new ReportingObserver((reports, observer) => resolve([reports, observer]),
                          {types: ['feature-policy-violation']}).observe();
  });
  await promise_rejects(t, 'SecurityError', navigator.xr.supportsSession('inline'),
                        "XR device access should not be allowed in this document.");
  const [reports, observer] = await report;
  check_report_format(reports, observer);
}, "XR Report Format");
    </script>
  </body>
</html>
